Modify bg jobs
There is a problem with the API. If you make 2 calls within 15 sec the call will fail. This is tricky when I need to make multiple calls. As the call for PDF could come in the middle of the calls towards the API, and that could trigger the throttle from the API. I've set the calls between the API requests to 30 sec, and PDF creation to 15 sec after API call. Hopefully this will prevent issues.